requirements.txt
# Core
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pydantic==2.9.2
beautifulsoup4==4.12.3
lxml==5.3.0

# AI/LLM
openai==1.51.0
httpx==0.27.2

# Utils
tqdm==4.66.5

# UI
streamlit==1.38.0
supabase==2.6.0
